fixes for recasting to get the code to work with latest gfortran compiler
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@3164 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/random_mod.F90
@@ -24,7 +24,7 @@ contains
     real(double), intent(out), optional :: getDoubleArray2d(:,:)
     integer(bigInt), save :: iSeed = 0 
     integer(bigInt) :: iseed_syncval
-    real :: r
+    real(double) :: r
     integer :: i, j
 
     !$OMP THREADPRIVATE (iSeed)
